+++ sentry-core/src/clientoptions.rs
@@ -12,6 +12,9 @@ use crate::{Integration, IntoDsn, TransportFactory};
 /// Type alias for before event/breadcrumb handlers.
 pub type BeforeCallback<T> = Arc<dyn Fn(T) -> Option<T> + Send + Sync>;
 
+/// Type alias for override sample rate callback.
+pub type OverrideSamplingRateCallback = Arc<dyn Fn(&Event<'static>) -> Option<f32> + Send + Sync>;
+
 /// The Session Mode of the SDK.
 ///
 /// Depending on the use-case, the SDK can be set to two different session modes:
@@ -220,6 +223,8 @@ pub struct ClientOptions {
     ///
     /// See [`before_breadcrumb`](method@ClientOptions::before_breadcrumb) for details.
     pub before_breadcrumb: Option<BeforeCallback<Breadcrumb>>,
+    /// Callback allowing for setting sampling rate based on user defined function.
+    pub override_sampling_rate: Option<OverrideSamplingRateCallback>,
     /// Callback that is executed for each Log being added.
     ///
     /// See [`before_send_log`](method@ClientOptions::before_send_log) for details.
@@ -580,6 +585,19 @@ impl ClientOptions {
         }
     }
 
+    /// Sets the [callback](field@ClientOptions::override_sampling_rate) that is used to override the
+    /// sampling rate.
+    #[inline]
+    pub fn override_sampling_rate(
+        self,
+        override_func: impl Fn(&Event<'static>) -> Option<f32> + Send + Sync + 'static,
+    ) -> Self {
+        Self {
+            override_sampling_rate: Some(Arc::new(override_func)),
+            ..self
+        }
+    }
+
     /// Sets the [callback](field@ClientOptions::before_send_log) that is executed before sending
     /// each log.
     #[cfg(feature = "logs")]
@@ -860,6 +878,7 @@ impl Default for ClientOptions {
             before_send_log: None,
             enable_metrics: true,
             before_send_metric: None,
+            override_sampling_rate: None,
         }
     }
 }